docker/context/gitlab-bot/requirements.txt
python_gitlab==1.4.0
urllib3==1.24
slackclient==1.3.0
setuptools==40.5.0
